extract rounding form field
