+++ src/components/Settings/Rounding/Rounding.css
@@ -10,8 +10,3 @@
   border-bottom-left-radius: 0;
   margin-right: 8px;
 }
-
-.Rounding-Container {
-  display: flex;
-  align-items: center;
-}
